+++ srcs/jobcontrol/mark_job_as_stopped.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   mark_job_as_stopped.c                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mpivet-p <mpivet-p@student.42.fr>          +#+  +:+       +#+        */
+/*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/08 04:04:30 by mpivet-p          #+#    #+#             */
-/*   Updated: 2020/02/08 04:04:39 by mpivet-p         ###   ########.fr       */
+/*   Updated: 2020/03/08 20:36:43 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "sh42.h"
 
-void	mark_job_as_stopped(t_job *job, int stopped)
+void	mark_job_as_stopped(t_job *job, int8_t stopped)
 {
 	t_lst	*ptr;
